mercilnew/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
sentence-transformers
chromadb
pandas
numpy
